ERC-6123: Move to Draft: Enlarged interface, streamlined reference implementation. (#25)
* Revised version of ERC-6123 including reference implementation for OTC-Derivatives and according Unit Tests

* Revised Doc, renamed SDC in SDCAbstract

* Revised Readme

* Changed Status to "DRAFT" - here we go again.

* Fixed some typos.

* Fixed some typos.

* Fixed typos. Fixed typo in file name.

* Removed unused variables.

* Reverted named change. Refactor rename SDCAbstract to SDC

* Changed the instruction to run out-of-the-box (clean checkout) wit

* Used view, where possible. Used local variable.

* Commenting our unused parameter.

* Fixed reference to image. Fixed blank lines after headlines.

* Fixed blank lines before code.

* Fixed headline (duplicate/wrong).

* Added note on NPM version numbers.

* Used ERC-20 in place of EIP-20

* Added abstract. Used Specification as headline.

* Fixed link to eip-20.md

* Minor fixed to README.md

* Minor fixes to README.md

* Simplified instruction to run unit tests, added package.json and hardhat.config.js

* Simplified instruction to run unit tests, added package.json and hardhat.config.js

* Removed duplicated resulted from manual merge with new ERC repo.

* Fixed link: eip-20.md -> erc-20.md

* Fixed Bug in SDCPledgedBalance.sol (afterTransfer), minor rewriting in markdown doc.

* Minor fixes (typos).

* Linking to eip instead of erc (see remark 7 ii on migrating pull requests to ERC repo).

* Linking to eip instead of erc (see remark 7 ii on migrating pull requests to ERC repo).

* Removed external link upon request.

* Update ERCS/erc-6123.md


* Update ERCS/erc-6123.md


* Update ERCS/erc-6123.md


---------
