+++ assets/erc-6123/doc/sdc_lifecycle_sequence_diagram.png
