+++ assets/erc-6123/hardhat.config.js
@@ -1,0 +1,6 @@
+/** @type import('hardhat/config').HardhatUserConfig */
+module.exports = {
+  solidity: "0.8.20",
+};
+require("@nomiclabs/hardhat-waffle"); 
+require('solidity-coverage');
